Bootstrap Statistic for a Super Sample

For each subsample in a given super sample, draws a normal subsample of the same size using the provided parameters, collects them into a new super sample, and computes a specified statistic on the resulting super sample.

Usage

one_bootstrap(x, hat_sigma, hat_gamma, geom, stat_fun)

Arguments

x

An object of class CSuperSample, representing the original super sample.

hat_sigma

The mean parameter for the normal distribution used in resampling.

hat_gamma

The covariance parameter for the normal distribution used in resampling.

geom

The geometry parameter to be passed to riemtan::rspdnorm.

stat_fun

A function to compute a statistic on the resulting CSuperSample object.

Value

The value returned by stat_fun when applied to the bootstrapped super sample.

Details

This function performs a parametric bootstrap by resampling each subsample in x according to the specified parameters, then aggregates the resampled data into a new CSuperSample and computes the desired statistic.