Compute p-values using bootstrap

Computes a bootstrap-based p-value and for a given super sample. The statistic used for the bootstrap can be specified via the stat_fun argument.

Usage

riem_anova(ss, stat_fun = log_wilks_lambda, den = 100)

Arguments

ss

An object of class CSuperSample.

stat_fun

A function to compute a statistic on the CSuperSample object (default: log_wilks_lambda).

den

The number of bootstrap samples to generate for estimating the p-value.

Value

numeric A bootstrap p-value.

Details

The function computes the statistic on the observed data and compares it to the distribution of statistics computed on bootstrapped samples.